requirements.txt
python-telegram-bot==20.6
matplotlib>=3.8.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.3
ta>=0.10.2
plotly>=5.18.0
seaborn>=0.12.2
python-dotenv>=1.0.0
sqlalchemy>=2.0.21
openpyxl>=3.1.2
aiohttp>=3.9.0
schedule>=1.2.1
pytest>=7.4.2
